imdb-ratings/requirements.txt
flask==2.2.0
requests==2.28.1
waitress==2.1.2
python-dotenv==0.20.0
pytest==7.1.2
pytest-cov==3.0.0